Add support for commands - additionally link build version info into release process
